requirements.txt
anyio==3.7.0
astunparse==1.6.3
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.2
graphql-core==3.2.3
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pydantic==1.10.8
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
strawberry-graphql==0.178.0
tomli==2.0.1
typing_extensions==4.6.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
